(#10) Fix property CRUD
Update to fix the states and make it more unified. But this may need to be cleaned up such that it can get the states from an API call instead of being hard-wired.

Updated the property views to be better formatted and fixed the calls such that you see state names instead of state ids.